Damage take system update

+++ Assets/Resources/Scripts/Characters/Stats/CharacterHealth.cs
@@ -1,6 +1,6 @@
 ﻿public class CharacterHealth : CharacterStat, IDamagable
 {
-    public virtual void TakeDamage(float damage)
+    public virtual void TakeDamage(int damage)
     {
         Decrease(damage);
     }
